requirements.txt
flask>=3.1
gunicorn>=23.0
psycopg2-binary>=2.9
pandas>=2.2
huggingface_hub>=1.0
